@oganyanrv Task2 variant 2
