+++ Oganyan.RV/Task2/Task2.vcxproj
@@ -146,7 +146,11 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClCompile Include="Source1.cpp" />
+    <ClCompile Include="Polynomial.cpp" />
+    <ClCompile Include="Polynomialtest.cpp" />
+  </ItemGroup>
+  <ItemGroup>
+    <ClInclude Include="Polynomial.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
