+++ Oganyan.RV/Task2/input.txt
@@ -1,0 +1,1 @@
+3 3 1
