+++ Oganyan.RV/Task2/output.txt
@@ -1,0 +1,1 @@
+2*x^4+3*x^2+4*x^1+5
